MAME 0.63

+++ src/sound/k053260.c
@@ -415,7 +415,7 @@ data8_t K053260_read( int chip, offs_t offset )
 
 		case 0x2e: /* read rom */
 			if ( ic->mode & 1 ) {
-				unsigned long offs = ic->channels[0].start + ( ic->channels[0].pos >> BASE_SHIFT ) + ( ic->channels[0].bank << 16 );
+				unsigned int offs = ic->channels[0].start + ( ic->channels[0].pos >> BASE_SHIFT ) + ( ic->channels[0].bank << 16 );
 
 				ic->channels[0].pos += ( 1 << 16 );
